프로그램 첨부파일 등록/수정 기능 구현
프로그램에 사진 첨부 파일 등록 기능 구현
수정모드일 때 기존 사진 불러오기 기능 구현
수정 시 파일 변경 여부 체크해서 변경되었을 경우 기존 파일 삭제하고 새로운 파일 등록하도록 설정

+++ src/main/java/egovframework/fourth/homework/service/impl/AttachserviceImpl.java
@@ -1,11 +1,15 @@
 package egovframework.fourth.homework.service.impl;
 
+import java.io.File;
+
 import javax.annotation.Resource;
 
 import org.egovframe.rte.fdl.cmmn.EgovAbstractServiceImpl;
+import org.egovframe.rte.fdl.property.EgovPropertyService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import egovframework.fourth.homework.service.AttachService;
 import egovframework.fourth.homework.service.AttachVO;
@@ -15,14 +19,38 @@ public class AttachserviceImpl extends EgovAbstractServiceImpl implements Attach
 	
 	private static final Logger log = LoggerFactory.getLogger(AttachserviceImpl.class);
 	
+    @Resource(name = "propertiesService")
+    private EgovPropertyService propertiesService;
+	
 	@Resource(name = "attachDAO")
 	private AttachDAO attachDAO;
 
-	// 첨부 파일 등록
+	// 첨부 파일 등록(프로그램)
 	@Override
-	public void createAttach(AttachVO vo) throws Exception {
-		attachDAO.insertAttach(vo);
-		log.info("INSERT 첨부파일({}) 등록 성공", vo.getIdx());
+	public void createProgramAttach(String programIdx, MultipartFile file) throws Exception {
+		if (file == null) return;
+        String baseDir = propertiesService.getString("file.upload.dir");
+        File uploadDir = new File(baseDir);
+        if (!uploadDir.exists()) uploadDir.mkdirs();
+        
+        String origName = file.getOriginalFilename();
+        String ext = origName.substring(origName.lastIndexOf('.'));
+        long size = file.getSize();
+        
+        // db에 질문 이미지 메타데이터 저장
+        AttachVO vo = new AttachVO();
+        vo.setProgramIdx(programIdx);
+        vo.setFileName(origName);
+        vo.setFilePath(baseDir);
+        vo.setFileSize(size);
+        vo.setExt(ext);
+        attachDAO.insertAttach(vo);
+        log.info("INSERT 프로그램({})에 이미지 첨부 파일({}) 등록 성공", vo.getProgramIdx(), vo.getIdx());
+        
+        // 실제 파일 물리 저장
+        File dest = new File(uploadDir, vo.getFileUuid() + ext);
+        file.transferTo(dest);
+        log.info("로컬 저장소에 파일 저장 완료! : {}", vo.getFileUuid() + vo.getExt());		
 	}
 
 	// 프로그램 idx로 첨부 이미지 파일 조회
@@ -44,16 +72,38 @@ public class AttachserviceImpl extends EgovAbstractServiceImpl implements Attach
 	// 첨부 파일 삭제
 	@Override
 	public void removeAttach(String idx) throws Exception {
-		attachDAO.deleteAttach(idx);
-		log.info("DELETE 첨부 파일({}) 삭제 완료", idx);
+		AttachVO vo = attachDAO.selectAttach(idx);
+		attachDAO.deleteAttach(idx); // DB에서 삭제
+		log.info("DELETE 질문 이미지({}) 삭제 완료", vo.getFileName());
+		// 물리 파일 삭제
+		File file = new File(vo.getFilePath(), vo.getFileUuid() + vo.getExt());
+		log.info("삭제되는 첨부파일 이름: {}", vo.getFileName());
+		if (file.exists()) {
+			file.delete();
+			log.info("로컬 저장소에서 파일 삭제 완료! : {}", vo.getFileUuid() + vo.getExt());
+		} else {
+			log.info("로컬 저장소에서 삭제할 파일이 존재하지 않음 : {}", vo.getFileUuid() + vo.getExt());
+		}
 	}
 	
 	// 프로그램에 있는 이미지 첨부파일 삭제
 	@Override
 	public void removeAttachByProgramIdx(String programIdx) throws Exception {
 		AttachVO vo = attachDAO.selectAttachByProgramIdx(programIdx);
-		attachDAO.deleteAttachByProgramIdx(programIdx);
+	    // 기존 파일이 없으면 리턴
+	    if (vo == null) return;
+		
+		attachDAO.deleteAttachByProgramIdx(programIdx); // DB에서 삭제
 		log.info("DELETE 프로그램({})에 대한 이미지 첨부 파일({}) 삭제 완료", programIdx, vo.getFileName());
+		// 물리 파일 삭제
+		File file = new File(vo.getFilePath(), vo.getFileUuid() + vo.getExt());
+		log.info("삭제될 프로그램({})에 소속된 삭제되는 첨부파일 이름: {}", programIdx, vo.getFileName());
+		if (file.exists()) {
+			file.delete();
+			log.info("로컬 저장소에서 파일 삭제 완료! : {}", vo.getFileUuid() + vo.getExt());
+		} else {
+			log.info("로컬 저장소에서 삭제할 파일이 존재하지 않음 : {}", vo.getFileUuid() + vo.getExt());
+		}
 	}
 
 }
